LayoutTests/platform/chromium/fast/forms/page-popup/page-popup-adjust-rect-expected.txt

Test for child elements of a datalist element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Move anchor element along the vertical axis.
PASS popupWindow.adjustWindowRect(30, 10, 30, 10).toString() is "Rect(0,0,30,10)"
PASS popupWindow.adjustWindowRect(30, 10, 30, 10).toString() is "Rect(0,0,30,10)"
PASS popupWindow.adjustWindowRect(30, 10, 30, 10).toString() is "Rect(0,10,30,10)"
PASS popupWindow.adjustWindowRect(30, 10, 30, 10).toString() is "Rect(0,20,30,10)"
PASS popupWindow.adjustWindowRect(30, 10, 30, 10).toString() is "Rect(0,30,30,10)"
PASS popupWindow.adjustWindowRect(30, 10, 30, 10).toString() is "Rect(0,40,30,10)"
PASS popupWindow.adjustWindowRect(30, 10, 30, 10).toString() is "Rect(0,50,30,10)"
PASS popupWindow.adjustWindowRect(30, 10, 30, 10).toString() is "Rect(0,40,30,10)"
PASS popupWindow.adjustWindowRect(30, 10, 30, 10).toString() is "Rect(0,50,30,10)"
PASS popupWindow.adjustWindowRect(30, 10, 30, 10).toString() is "Rect(0,50,30,10)"
PASS popupWindow.adjustWindowRect(30, 10, 30, 10).toString() is "Rect(0,50,30,10)"
Move anchor element along the horizontal axis.
PASS popupWindow.adjustWindowRect(30, 10, 30, 10).toString() is "Rect(0,10,30,10)"
PASS popupWindow.adjustWindowRect(30, 10, 30, 10).toString() is "Rect(0,10,30,10)"
PASS popupWindow.adjustWindowRect(30, 10, 30, 10).toString() is "Rect(0,10,30,10)"
PASS popupWindow.adjustWindowRect(30, 10, 30, 10).toString() is "Rect(0,10,30,10)"
PASS popupWindow.adjustWindowRect(30, 10, 30, 10).toString() is "Rect(0,10,30,10)"
PASS popupWindow.adjustWindowRect(30, 10, 30, 10).toString() is "Rect(10,10,30,10)"
PASS popupWindow.adjustWindowRect(30, 10, 30, 10).toString() is "Rect(20,10,30,10)"
PASS popupWindow.adjustWindowRect(30, 10, 30, 10).toString() is "Rect(30,10,30,10)"
PASS popupWindow.adjustWindowRect(30, 10, 30, 10).toString() is "Rect(40,10,30,10)"
PASS popupWindow.adjustWindowRect(30, 10, 30, 10).toString() is "Rect(50,10,30,10)"
PASS popupWindow.adjustWindowRect(30, 10, 30, 10).toString() is "Rect(50,10,30,10)"
PASS popupWindow.adjustWindowRect(30, 10, 30, 10).toString() is "Rect(50,10,30,10)"
PASS popupWindow.adjustWindowRect(30, 10, 30, 10).toString() is "Rect(50,10,30,10)"
PASS popupWindow.adjustWindowRect(30, 10, 30, 10).toString() is "Rect(50,10,30,10)"
PASS popupWindow.adjustWindowRect(30, 10, 30, 10).toString() is "Rect(50,10,30,10)"
Move anchor element along the vertical axis. Open popup that is too tall.
PASS popupWindow.adjustWindowRect(30, 1000, 30, 10).toString() is "Rect(0,0,30,60)"
PASS popupWindow.adjustWindowRect(30, 1000, 30, 10).toString() is "Rect(0,0,30,60)"
PASS popupWindow.adjustWindowRect(30, 1000, 30, 10).toString() is "Rect(0,10,30,50)"
PASS popupWindow.adjustWindowRect(30, 1000, 30, 10).toString() is "Rect(0,20,30,40)"
PASS popupWindow.adjustWindowRect(30, 1000, 30, 10).toString() is "Rect(0,30,30,30)"
PASS popupWindow.adjustWindowRect(30, 1000, 30, 10).toString() is "Rect(0,0,30,30)"
PASS popupWindow.adjustWindowRect(30, 1000, 30, 10).toString() is "Rect(0,0,30,40)"
PASS popupWindow.adjustWindowRect(30, 1000, 30, 10).toString() is "Rect(0,0,30,50)"
PASS popupWindow.adjustWindowRect(30, 1000, 30, 10).toString() is "Rect(0,0,30,60)"
PASS popupWindow.adjustWindowRect(30, 1000, 30, 10).toString() is "Rect(0,0,30,60)"
PASS popupWindow.adjustWindowRect(30, 1000, 30, 10).toString() is "Rect(0,0,30,60)"
Move anchor element along the vertical axis. Open popup that is too tall and won't shrink.
PASS popupWindow.adjustWindowRect(30, 1000, 30, 1000).toString() is "Rect(0,0,30,1000)"
PASS popupWindow.adjustWindowRect(30, 1000, 30, 1000).toString() is "Rect(0,0,30,1000)"
PASS popupWindow.adjustWindowRect(30, 1000, 30, 1000).toString() is "Rect(0,0,30,1000)"
PASS popupWindow.adjustWindowRect(30, 1000, 30, 1000).toString() is "Rect(0,0,30,1000)"
PASS popupWindow.adjustWindowRect(30, 1000, 30, 1000).toString() is "Rect(0,0,30,1000)"
PASS popupWindow.adjustWindowRect(30, 1000, 30, 1000).toString() is "Rect(0,0,30,1000)"
PASS popupWindow.adjustWindowRect(30, 1000, 30, 1000).toString() is "Rect(0,0,30,1000)"
PASS popupWindow.adjustWindowRect(30, 1000, 30, 1000).toString() is "Rect(0,0,30,1000)"
PASS popupWindow.adjustWindowRect(30, 1000, 30, 1000).toString() is "Rect(0,0,30,1000)"
PASS popupWindow.adjustWindowRect(30, 1000, 30, 1000).toString() is "Rect(0,0,30,1000)"
PASS popupWindow.adjustWindowRect(30, 1000, 30, 1000).toString() is "Rect(0,0,30,1000)"
Grow popup width.
PASS popupWindow.adjustWindowRect(30, 10, 30, 10).toString() is "Rect(30,10,30,10)"
PASS popupWindow.adjustWindowRect(40, 10, 40, 10).toString() is "Rect(30,10,40,10)"
PASS popupWindow.adjustWindowRect(50, 10, 50, 10).toString() is "Rect(30,10,50,10)"
PASS popupWindow.adjustWindowRect(60, 10, 60, 10).toString() is "Rect(20,10,60,10)"
PASS popupWindow.adjustWindowRect(70, 10, 70, 10).toString() is "Rect(10,10,70,10)"
PASS popupWindow.adjustWindowRect(80, 10, 80, 10).toString() is "Rect(0,10,80,10)"
PASS popupWindow.adjustWindowRect(90, 10, 90, 10).toString() is "Rect(0,10,90,10)"
PASS successfullyParsed is true

TEST COMPLETE
 
